requirements.txt
websockets
httpx[http2]
pystray
pillow
python-vlc
